Move ObjectType to __init_subclass__
